Moved _prepare_data_in_groups to xgboostextension.util module.
